requirements.txt
dm_control==1.0.19
dm_env==1.6
glfw==2.7.0
gymnasium==0.29.1
hydra-core==1.3.2
lxml==4.9.1
lxml==5.2.2
matlabengine==24.1
matlabengine.egg==info
matplotlib==3.7.0
matplotlib==3.8.0
mplcursors==0.5.2
mujoco==3.1.6
numpy==1.24.3
numpy==1.23.5
omegaconf==2.3.0
opencv_python==4.10.0.82
pynput==1.7.7
rich==13.7.1
Shimmy==1.3.0
stable_baselines3==2.3.2
sympy==1.11.1
sympy==1.12.1
termcolor==2.4.0
torch==1.12.1
torch==2.3.0
tqdm==4.64.1
tqdm==4.66.4
tyro==0.8.4
wandb==0.17.7
